+++ gcloud.go
@@ -1,11 +1,8 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"golang.org/x/oauth2"
-	"os/exec"
 	"time"
 )
 
@@ -13,39 +10,24 @@ type gcloudTokenSource struct {
 	cfg *gcloudConfig
 }
 
-type gcloudCredential struct {
-	AccessToken string    `json:"access_token"`
-	IdToken     string    `json:"id_token"`
-	TokenExpiry time.Time `json:"token_expiry"`
-}
-
 type gcloudConfig struct {
-	Credential gcloudCredential `json:"credential"`
-	Core       struct {
+	Credential struct {
+		AccessToken string    `json:"access_token"`
+		IdToken     string    `json:"id_token"`
+		TokenExpiry time.Time `json:"token_expiry"`
+	} `json:"credential"`
+	Core struct {
 		Account string `json:"account"`
 	} `json:"core"`
 }
 
-func getGcloudConfig(account string) (*gcloudConfig, error) {
-	var buf bytes.Buffer
-	args := []string{"config", "config-helper", "--format=json"}
-	if account != "" {
-		args = append(args, "--account="+account)
-	}
-
-	cmd := exec.Command("gcloud", args...)
-	cmd.Stdout = &buf
-	err := cmd.Run()
+func GcloudTokenSource(account string) (oauth2.TokenSource, error) {
+	cfg, err := fetchGcloudConfig(account)
 	if err != nil {
 		return nil, err
 	}
 
-	var parsed gcloudConfig
-	err = json.Unmarshal(buf.Bytes(), &parsed)
-	if err != nil {
-		return nil, err
-	}
-	return &parsed, nil
+	return &gcloudTokenSource{cfg}, nil
 }
 
 func (gts *gcloudTokenSource) Token() (*oauth2.Token, error) {
@@ -53,15 +35,6 @@ func (gts *gcloudTokenSource) Token() (*oauth2.Token, error) {
 		AccessToken: gts.cfg.Credential.AccessToken,
 		Expiry:      gts.cfg.Credential.TokenExpiry,
 	}, nil
-}
-
-func newGcloudTokenSource(account string) (oauth2.TokenSource, error) {
-	cfg, err := getGcloudConfig(account)
-	if err != nil {
-		return nil, err
-	}
-
-	return &gcloudTokenSource{cfg}, nil
 }
 
 func (gts *gcloudTokenSource) Email() (string, error) {
