+++ metadata.go
@@ -12,28 +12,21 @@ type metadataTokenSource struct {
 	account string
 }
 
-func newMetadataTokenSource(account string) (*metadataTokenSource, error) {
+func MetadataTokenSource(account string) (*metadataTokenSource, error) {
 	return &metadataTokenSource{account: account}, nil
 }
 
-func newMetadataTokenSourceDefault() (*metadataTokenSource, error) {
-	return newMetadataTokenSource("")
+func MetadataTokenSourceDefault() (*metadataTokenSource, error) {
+	return MetadataTokenSource("")
 }
 
-func orDefault(v string, def string) string {
-	if v == "" {
-		return def
-	}
-	return v
-}
-
-func (gts *metadataTokenSource) Token() (*oauth2.Token, error) {
-	tokenSource := google.ComputeTokenSource(gts.account)
+func (mts *metadataTokenSource) Token() (*oauth2.Token, error) {
+	tokenSource := google.ComputeTokenSource(mts.account)
 	return tokenSource.Token()
 }
 
-func (gts *metadataTokenSource) AccessToken(ctx context.Context, scopes ...string) (string, error) {
-	tokenSource := google.ComputeTokenSource(gts.account, scopes...)
+func (mts *metadataTokenSource) AccessToken(ctx context.Context, scopes ...string) (string, error) {
+	tokenSource := google.ComputeTokenSource(mts.account, scopes...)
 	token, err := tokenSource.Token()
 	if err != nil {
 		return "", err
@@ -41,18 +34,18 @@ func (gts *metadataTokenSource) AccessToken(ctx context.Context, scopes ...strin
 	return token.AccessToken, nil
 }
 
-func (gts *metadataTokenSource) IDToken(ctx context.Context, audience string) (string, error) {
+func (mts *metadataTokenSource) IDToken(ctx context.Context, audience string) (string, error) {
 	params := make(url.Values)
 	params.Set("audience", audience)
-	tokenString, err := metadata.Get("instance/service-accounts/" + orDefault(gts.account, "default") + "/identity?" + params.Encode())
+	tokenString, err := metadata.Get("instance/service-accounts/" + orDefault(mts.account, "default") + "/identity?" + params.Encode())
 	if err != nil {
 		return "", err
 	}
 	return tokenString, err
 }
 
-func (gts *metadataTokenSource) Email(ctx context.Context, audience string) (string, error) {
-	tokenString, err := metadata.Get("instance/service-accounts/" + orDefault(gts.account, "default") + "/email")
+func (mts *metadataTokenSource) Email(ctx context.Context, audience string) (string, error) {
+	tokenString, err := metadata.Get("instance/service-accounts/" + orDefault(mts.account, "default") + "/email")
 	if err != nil {
 		return "", err
 	}
